+++ src/backend/joanie/core/admin.py
@@ -119,6 +119,7 @@ class OrderAdmin(admin.ModelAdmin):
     """Admin class for the Order model"""
 
     list_display = ("owner", "product", "state")
+    readonly_fields = ("price",)
 
 
 @admin.register(models.Enrollment)